Revert "Revert "Rename: Move Header component""
This reverts commit 868be9c5604d28847a9007de1868e9cbe0ab61ff.
