+++ src/app/auth/signin/page.tsx
@@ -5,8 +5,7 @@ import { useRouter } from 'next/navigation';
 import { useEffect } from 'react';
 
 import { SigninForm } from '@/components/auth';
-import Header from '@/components/common/Header';
-import HeaderNav from '@/components/common/HeaderNav';
+import { Header, HeaderNav } from '@/components/common/header';
 
 import authRequest from '@/api/authRequest';
