+++ src/app/mypage/page.tsx
@@ -3,8 +3,7 @@
 import { useRouter } from 'next/navigation';
 import { useEffect, useState } from 'react';
 
-import Header from '@/components/common/Header';
-import HeaderNav from '@/components/common/HeaderNav';
+import { Header, HeaderNav } from '@/components/common/header';
 import { HistoryButton, LogoutButton } from '@/components/mypage';
 
 import authRequest from '../../api/authRequest';
