+++ src/app/page.tsx
@@ -3,9 +3,8 @@
 'use client';
 import React, { useState } from 'react';
 
-import HeaderNav from '@/components/common/HeaderNav';
+import { Header, HeaderNav } from '@/components/common/header';
 
-import Header from '../components/common/Header';
 import Carousel from '../components/main/Carousel';
 import IconList from '../components/main/IconList';
 import Navigation from '../components/main/Navigation';
@@ -335,9 +334,7 @@ const HomePage = () => {
   return (
     <div className='bg-white pt-[3rem]'>
       <Header>
-        <HeaderNav showLogo showCart showMyPage>
-          {' '}
-        </HeaderNav>
+        <HeaderNav showLogo showCart showMyPage />
       </Header>
       <Carousel images={carouselImages} />
       <IconList icons={mainIcons} />
