+++ src/components/common/header/index.ts
@@ -1,0 +1,4 @@
+import Header from './Header';
+import HeaderNav from './HeaderNav';
+
+export { Header, HeaderNav };
